# Speed up `readtable`