--- a/src/XLSX.jl
+++ b/src/XLSX.jl
@@ -54,7 +54,6 @@ include("cellformats.jl")
 include("conditional-format-helpers.jl") # must load before conditional-formats.jl
 include("conditional-formats.jl")
 include("write.jl")
-include("fileArray.jl")
 
 PCT.@setup_workload begin
     # Putting some things in `@setup_workload` instead of `@compile_workload` can reduce the size of the