This PR is motivated by observations by @chris-b1 in comments to #316.

I've eeked out a little more performance from `readtable`. It is now within a small margin of v0.10.4, while pretty much everything else remains much faster (per the comparisons in #315 and #318).

Here are some readtable comparisons, based on a 650k row, 16 column Excel file:

**This PR**
```
    readtable to df      48.267293 seconds (2.07 G allocations: 126.013 GiB, 44.81% gc time, 9 lock conflicts)
    writetable to file   29.079795 seconds (327.61 M allocations: 21.600 GiB, 26.77% gc time, 13 lock conflicts)
```

**v0.10.4**
```
    readtable to df      32.743151 seconds (376.26 M allocations: 26.203 GiB, 23.66% gc time)
    writetable to file   30.392381 seconds (139.93 M allocations: 6.012 GiB, 3.89% gc time)
```